Split out QSharpCore lib
