+++ src/QirRuntime/lib/CMakeLists.txt
@@ -1,4 +1,4 @@
 add_subdirectory(QIR)
+add_subdirectory(QSharpCore)
 add_subdirectory(Simulators)
 add_subdirectory(Tracer)
-add_subdirectory(qdk)
