+++ src/QirRuntime/lib/QSharpCore/.clang-tidy
@@ -1,0 +1,1 @@
+Checks: '-*,bugprone-*'
