--- src/QirRuntime/lib/qdk/CMakeLists.txt
@@ -1,28 +1,0 @@
-#===============================================================================
-# Produce the qdk dynamic library
-#
-add_library(qdk SHARED
-  qdk.cpp
-  $<TARGET_OBJECTS:qir-rt-support-obj>
-  $<TARGET_OBJECTS:qir-qis-support-obj>
-  $<TARGET_OBJECTS:simulators-obj>
-)
-
-target_link_libraries(qdk
-  ${QIR_BRIDGE_UTILITY_LIB}
-  ${QIR_BRIDGE_QIS_UTILITY_LIB}
-  ${CMAKE_DL_LIBS}
-)
-
-target_include_directories(qdk PUBLIC 
-   ${public_includes}
-)
-
-set_property(TARGET qdk PROPERTY POSITION_INDEPENDENT_CODE ON)
-
-install(TARGETS qdk
-  RUNTIME DESTINATION "${CMAKE_BINARY_DIR}/bin" 
-  LIBRARY DESTINATION "${CMAKE_BINARY_DIR}/bin"
-)
-
-
