+++ src/QirRuntime/test/QIR-dynamic/CMakeLists.txt
@@ -8,15 +8,9 @@ if (CMAKE_BUILD_TYPE STREQUAL "Debug")
   )
 endif()
 
-if (WIN32)
-  set(TEST_FILES
-    qir-test-random-win
-  )
-else()
-  set(TEST_FILES
-    qir-test-random-lnx
-  )
-endif()
+set(TEST_FILES
+  qir-test-random
+)
 
 foreach(file ${TEST_FILES})
   compile_from_qir(${file} "") # don't create a target per file
@@ -26,8 +20,8 @@ endforeach()
 add_custom_target(qir_dynamic_test_lib DEPENDS ${QIR_TESTS_LIBS})
 
 #==============================================================================
-# This executable target links test code against the dynamic QDK library rather than the explicit
-# static QIR/RT libs (qdk will statically link in the bridge via transitivity of target_link_libraries).
+# This executable target links test code against the dynamic QIR and QSharpCore libraries rather than the explicit
+# static QIR/RT libs (qir will statically link in the bridge via transitivity of target_link_libraries).
 #
 add_executable(qir-dynamic-tests
   qir-driver.cpp
@@ -35,7 +29,8 @@ add_executable(qir-dynamic-tests
 
 target_link_libraries(qir-dynamic-tests PUBLIC
   ${QIR_TESTS_LIBS}
-  qdk
+  qir
+  QSharpCore
 )
 
 target_include_directories(qir-dynamic-tests PUBLIC
