+++ src/QirRuntime/test/unittests/CMakeLists.txt
@@ -11,6 +11,7 @@ add_executable(qir-runtime-unittests
 target_link_libraries(qir-runtime-unittests PUBLIC
   qir-rt-support
   qir-qis-support
+  qsharp-core-qis-support
   simulators
   tracer
 )
@@ -18,8 +19,9 @@ target_link_libraries(qir-runtime-unittests PUBLIC
 target_include_directories(qir-runtime-unittests PUBLIC 
   "${test_includes}"
    ${public_includes}
-  "${PROJECT_SOURCE_DIR}/lib/QIR"
-  "${PROJECT_SOURCE_DIR}/lib/Tracer"
+   "${PROJECT_SOURCE_DIR}/lib/QIR"
+   "${PROJECT_SOURCE_DIR}/lib/QSharpCore"
+   "${PROJECT_SOURCE_DIR}/lib/Tracer"
 )
 install(TARGETS qir-runtime-unittests RUNTIME DESTINATION "${CMAKE_BINARY_DIR}/bin")
 add_unit_test(qir-runtime-unittests)
